system_win32: Moved the Curl_load_library() function into a seperate module
This allows the curl tool to only share this function rather than all
the initialisation stuff that libcurl needs from system_win32.
